juice

+++ native/src/init/rootdir.cpp
@@ -51,23 +51,20 @@ static void patch_init_rc(const char *src, const char *dest, const char *tmp_dir
     rc_list.clear();
 
     // Inject Magisk rc scripts
-    char pfd_svc[16], ls_svc[16];
-    gen_rand_str(pfd_svc, sizeof(pfd_svc));
-    gen_rand_str(ls_svc, sizeof(ls_svc));
-    LOGD("Inject magisk services: [%s] [%s]\n", pfd_svc, ls_svc);
-    fprintf(rc, MAGISK_RC, tmp_dir, pfd_svc, ls_svc);
+    LOGD("Inject magisk services\n");
+    fprintf(rc, MAGISK_RC, tmp_dir);
 
     fclose(rc);
     clone_attr(src, dest);
 }
 
-static void load_overlay_rc(const char *overlay) {
+static void load_overlay_rc(const char *overlay, bool should_unlink = true) {
     auto dir = open_dir(overlay);
     if (!dir) return;
 
     int dfd = dirfd(dir.get());
     // Do not allow overwrite init.rc
-    unlinkat(dfd, "init.rc", 0);
+    if (should_unlink) unlinkat(dfd, "init.rc", 0);
 
     // '/' + name + '\0'
     char buf[NAME_MAX + 2];
@@ -84,7 +81,7 @@ static void load_overlay_rc(const char *overlay) {
             int rc = xopenat(dfd, entry->d_name, O_RDONLY | O_CLOEXEC);
             rc_list.push_back(full_read(rc));
             close(rc);
-            unlinkat(dfd, entry->d_name, 0);
+            if (should_unlink) unlinkat(dfd, entry->d_name, 0);
         }
     }
 }
@@ -192,7 +189,7 @@ void MagiskInit::patch_ro_root() {
     if (access("/sbin", F_OK) == 0) {
         tmp_dir = "/sbin";
     } else {
-        char buf[8];
+        char buf[16];
         gen_rand_str(buf, sizeof(buf));
         tmp_dir = "/dev/"s + buf;
         xmkdir(tmp_dir.data(), 0);
@@ -228,6 +225,7 @@ void MagiskInit::patch_ro_root() {
 #endif
 
     load_overlay_rc(ROOTOVL);
+    load_overlay_rc(INTLROOT "/early-mount.d/initrc.d", false);
     if (access(ROOTOVL "/sbin", F_OK) == 0) {
         // Move files in overlay.d/sbin into tmp_dir
         mv_path(ROOTOVL "/sbin", ".");
@@ -267,8 +265,7 @@ void RootFSInit::prepare() {
     rename(backup_init(), "/init");
 }
 
-#define PRE_TMPSRC "/magisk"
-#define PRE_TMPDIR PRE_TMPSRC "/tmp"
+#define PRE_TMPDIR "/magisk-tmp"
 
 void MagiskInit::patch_rw_root() {
     mount_list.emplace_back("/data");
@@ -283,20 +280,22 @@ void MagiskInit::patch_rw_root() {
     rm_rf("/data/overlay.d");
     rm_rf("/.backup");
 
-    // Patch init.rc
-    patch_init_rc("/init.rc", "/init.p.rc", "/sbin");
-    rename("/init.p.rc", "/init.rc");
-
     bool treble;
     {
         auto init = mmap_data("/init");
         treble = init.contains(SPLIT_PLAT_CIL);
     }
 
-    xmkdir(PRE_TMPSRC, 0);
-    xmount("tmpfs", PRE_TMPSRC, "tmpfs", 0, "mode=755");
     xmkdir(PRE_TMPDIR, 0);
     setup_tmp(PRE_TMPDIR);
+
+    // Handle custom rc script
+    load_overlay_rc(PRE_TMPDIR "/" MIRRDIR "/early-mount/initrc.d", false);
+
+    // Patch init.rc
+    patch_init_rc("/init.rc", "/init.p.rc", "/sbin");
+    rename("/init.p.rc", "/init.rc");
+
     chdir(PRE_TMPDIR);
 
     // Extract magisk
@@ -322,12 +321,10 @@ int magisk_proxy_main(int argc, char *argv[]) {
     unlink("/sbin/magisk");
 
     // Move tmpfs to /sbin
-    // make parent private before MS_MOVE
-    xmount(nullptr, PRE_TMPSRC, nullptr, MS_PRIVATE, nullptr);
-    xmount(PRE_TMPDIR, "/sbin", nullptr, MS_MOVE, nullptr);
-    xumount2(PRE_TMPSRC, MNT_DETACH);
+    // For some reason MS_MOVE won't work, as a workaround bind mount then unmount
+    xmount(PRE_TMPDIR, "/sbin", nullptr, MS_BIND | MS_REC, nullptr);
+    xumount2(PRE_TMPDIR, MNT_DETACH);
     rmdir(PRE_TMPDIR);
-    rmdir(PRE_TMPSRC);
 
     // Create symlinks pointing back to /root
     recreate_sbin("/root", false);
